+++ packages/cs-shared/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `UBEntity.isManyManyRef` property added. `true` in case this is many-to-many relation table for storing values of attributes of "Many" type.
 
 ### Changed